+++ src/lib/game/config.ts
@@ -8,3 +8,14 @@ export const enum Const {
 export let _debugLagK = 0;
 
 export const setDebugLagK = (a: number) => _debugLagK = a;
+
+export const GAME_CFG = {
+    npc: {
+        period: 10,
+        // max: 8
+        max: 0
+    },
+    items: {
+        lifetime: 10 * Const.NetFq
+    }
+};